refactor(api): move logModeration to moderationEventService
