+++ api/src/index.ts
@@ -29,8 +29,8 @@ import CampaignController from "./controllers/campaign";
 import IframeController from "./controllers/iframe";
 import ImportController from "./controllers/import";
 import MissionController from "./controllers/mission";
-import ModerationController from "./controllers/moderation";
-import ModerationEventController from "./controllers/moderation-event";
+import ModerationController from "./controllers/moderation/moderation";
+import ModerationEventController from "./controllers/moderation/moderation-event";
 import OrganizationController from "./controllers/organization";
 import PublisherController from "./controllers/publisher";
 import RedirectController from "./controllers/redirect";
